project/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
pydantic>=2.11.0
python-multipart>=0.0.20
faster-whisper>=1.1.0
paho-mqtt>=2.1.0
httpx>=0.28.0
PyYAML>=6.0.2
numpy>=1.26.0
openwakeword>=0.6.0
onnxruntime>=1.17.0
